Update the node versions.
There's a bug affecting iojs-3.3, node-4.0 and node-4.1 that stops HTTP
request streams emitting data in some scenarios. This causes many tests
in the suite to fail when closing the socket, delaying the duration of
each test.

We're going to test on major releases that are free of this issue and
resume testing when the issue is fixed, which is expected in 5.0.